Chore refactor file structure
