+++ frontend/public/svg/Close.tsx
