+++ frontend/public/svg/Copy.tsx
