+++ frontend/public/svg/Library.tsx
