+++ frontend/public/svg/Menu.tsx
