+++ frontend/public/svg/Moon.tsx
