+++ frontend/public/svg/Sun.tsx
