+++ frontend/public/svg/Upright.tsx
