+++ frontend/public/svg/Verified.tsx
