+++ frontend/public/svg/WarnBadge.tsx
