+++ frontend/src/app/components/internal/Footer.tsx
