+++ frontend/src/app/components/internal/Header.tsx
@@ -1,11 +1,11 @@
 "use client";
 import { useEffect, useRef } from "react";
 import { useAccount } from "@starknet-react/core";
-import AddressBar from "./AddressBar";
-import ThemeSwitch from "./ThemeSwitch";
-import Transactions from "./Transactions";
-import ConnectButton from "./Connect";
-import useTheme from "@/app/hooks/useTheme";
+import AddressBar from "../lib/AddressBar";
+import ThemeSwitch from "./util/ThemeSwitch";
+import Transactions from "../lib/Transactions";
+import ConnectButton from "../lib/Connect";
+import useTheme from "@/app/components/internal/hooks/useTheme";
 import Link from "next/link";
 
 const Header = () => {
