+++ frontend/src/app/components/internal/helpers/constant.ts
