+++ frontend/src/app/components/lib/AddressBar.tsx
@@ -5,11 +5,11 @@ import {
   useStarkProfile,
 } from "@starknet-react/core";
 import Blockies from "react-blockies";
-import AccountBalance from "../AccountBalance";
-import GenericModal from "../GenericModal";
-import Close from "svg/Close";
+import AccountBalance from "./AccountBalance";
+import GenericModal from "../internal/util/GenericModal";
+import Close from "public/svg/Close";
 import { useEffect, useState } from "react";
-import CopyButton from "../util/CopyButton";
+import CopyButton from "../internal/util/CopyButton";
 
 const UserModal = () => {
   const { address } = useAccount();
