+++ frontend/src/app/layout.tsx
@@ -1,7 +1,7 @@
 import type { Metadata } from "next";
 import "./globals.css";
 import { StarknetProvider } from "~/StarknetProvider";
-import Footer from "./components/ui_components/Footer";
+import Footer from "./components/internal/Footer";
 
 export const metadata: Metadata = {
   title: "Starknet Scaffold",
