--- frontend/src/app/types/index.ts
@@ -1,15 +1,0 @@
-export interface WikipediaResource {
-  id: string;
-  name: string;
-  description: string;
-  status: string;
-  url: string;
-  category: string;
-}
-
-export interface AddressBookResource {
-  id: string;
-  name: string;
-  description: string;
-  address: string;
-}
