--- frontend/src/app/utils/currency.ts
@@ -1,4 +1,0 @@
-export function formatCurrency(currency: number) {
-  let amount = currency / 1e18;
-  return amount;
-}
