--- frontend/src/app/utils/date.ts
@@ -1,11 +1,0 @@
-export function formatDate(isoString: string): string {
-  const options: Intl.DateTimeFormatOptions = {
-    day: "numeric",
-    month: "2-digit",
-    year: "numeric",
-    hour: "2-digit",
-    minute: "2-digit",
-  };
-  const date = new Date(isoString);
-  return date.toLocaleDateString("en-US", options);
-}
